// ACP-Checkout-Gateway/src/api.ts
export interface components {
    schemas: {
// ...
        FulfillmentDetails: {
            name?: string;
            phone_number?: string;
            /** Format: email */
            email?: string;
            address?: components["schemas"]["Address"];
        };
        Buyer: {
            first_name: string;
            last_name: string;
            /** Format: email */
            email: string;
            phone_number?: string;
        };
// ...
        LineItem: {
            id: string;
            item: components["schemas"]["Item"];
            base_amount: number;
            discount: number;
            subtotal: number;
            tax: number;
            total: number;
            name?: string;
            description?: string;
            images?: string[];
            unit_amount?: number;
            disclosures?: components["schemas"]["Disclosure"][];
            custom_attributes?: components["schemas"]["CustomAttribute"][];
            marketplace_seller_details?: components["schemas"]["MarketplaceSellerDetails"];
        };
        Total: {
            /** @enum {string} */
            type: "items_base_amount" | "items_discount" | "subtotal" | "discount" | "fulfillment" | "tax" | "fee" | "total";
            display_text: string;
            amount: number;
            description?: string;
        };
        FulfillmentOptionShipping: {
            /** @constant */
            type: "shipping";
            id: string;
            title: string;
            subtitle?: string;
            carrier?: string;
            /** Format: date-time */
            earliest_delivery_time?: string;
            /** Format: date-time */
            latest_delivery_time?: string;
            subtotal: number;
            tax: number;
            total: number;
        };
        FulfillmentOptionDigital: {
            /** @constant */
            type: "digital";
            id: string;
            title: string;
            subtitle?: string;
            subtotal: number;
            tax: number;
            total: number;
        };
        SelectedShippingFulfillmentOption: {
            option_id: string;
            item_ids: string[];
        };
        SelectedDigitalFulfillmentOption: {
            option_id: string;
            item_ids: string[];
        };
        SelectedFulfillmentOption: {
            /** @enum {string} */
            type: "shipping" | "digital";
            shipping?: components["schemas"]["SelectedShippingFulfillmentOption"];
            digital?: components["schemas"]["SelectedDigitalFulfillmentOption"];
        } & ({
            /** @constant */
            type?: "shipping";
        } | {
            /** @constant */
            type?: "digital";
        });
        MessageInfo: {
            /** @constant */
            type: "info";
            /** @description RFC 9535 JSONPath */
            param?: string;
            /** @enum {string} */
            content_type: "plain" | "markdown";
            content: string;
        };
// ...
        Order: {
            id: string;
            checkout_session_id: string;
            /** Format: uri */
            permalink_url: string;
        };
// ...
        CheckoutSession: components["schemas"]["CheckoutSessionBase"];
        CheckoutSessionWithOrder: components["schemas"]["CheckoutSessionBase"] & {
            order: components["schemas"]["Order"];
        };
// ...
    };
    responses: never;
    parameters: {
        Authorization: string;
        ContentType: string;
        AcceptLanguage: string;
        UserAgent: string;
        IdempotencyKey: string;
        RequestId: string;
        Signature: string;
        Timestamp: string;
        APIVersion: string;
    };
    requestBodies: never;
    headers: never;
    pathItems: never;
}